Create header for timeframe workspace, move all tools to the header
