+++ src/index.css
@@ -1,5 +1,5 @@
 :root {
-  font-family: Saira Condensed;
+  font-family: Roboto Condensed;
   line-height: 1.5;
   font-weight: 400;
